chore: improve debugging
Add more debugging so that its easier to understand what is happening.

Default debugging setting to that the DEBUG environment if set so that
the developer doesn't have to set it manually in settings when running a
debug session.

Also enable npm-groovy-lint debugging which is key to understanding
issues. While this was previous set for debugging sessions in the
environment the setting would always override it.